requirements.txt
pytest~=7.1.2
selenium~=4.4.3
requests~=2.28.1
pandas~=0.24.2
styleframe~=2.0.5
chardet~=3.0.4
thrift~=0.11.0
pillow~=6.0.0
numpy~=1.16.4
xlrd~=1.2.0
xlutils~=2.0.0
pymysql~=0.9.2
python-jenkins~=1.5.0
enum34~=1.1.10
appium-python-client~=2.6.1
redis~=3.3.11
pika~=1.1.0